+++ src/main/java/ek/kinoxp/Repository/MovieRepository.java
@@ -3,7 +3,6 @@ package ek.kinoxp.Repository;
 import ek.kinoxp.Model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer>
-{
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 
 }
