+++ src/main/java/ek/kinoxp/Repository/ShowRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface ShowRepository extends JpaRepository<Show, Integer>
+public interface ShowRepository extends JpaRepository<Show, Long>
 {
     List<Show> findByStartTimeBetween(LocalDateTime start, LocalDateTime end);
     List<Show> findByStartTimeAfterOrderByStartTimeAsc(LocalDateTime t);
